started moving rooms reducer
