+++ src/features/input/ChatInput.jsx
@@ -12,7 +12,6 @@ import {
 } from "../../selectors/selectors.js";
 import styled from "styled-components";
 import theme from "../../constants/theme.js";
-import { sendRoomMessage } from "../room/roomActions.js";
 import { updateMessage } from "../message/messageActions.js";
 import { isMobile } from "../../constants/browserInfo.js";
 import {
@@ -24,6 +23,7 @@ import {
 	selectUpInEmoticonPicker,
 	showEmoticonPicker
 } from "../emoticon/emoticonPickerActions";
+import { sendRoomMessage } from "../room/roomsSlice";
 
 const removeNewlines = text => text.replace(/([\n\r])+/, "");
